Move tooltip from writer to parser
